python/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pyjwt